+++ helper-scripts/src/stackManagementTool/selectTool/selectTool.ts
@@ -1,7 +1,6 @@
 import { echo } from "zx";
-import { whichDeletionTool } from "../delete/utils/whichDeletionTool.js";
 import { deployDevStack } from "../deploy/deployDevStack.js";
-import { askWhichTool, welcomeMessage } from "./prompts.js";
+import { askWhichTool, welcomeMessage, whichDeletionTool } from "./prompts.js";
 
 export const selectTool = async (): Promise<void> => {
   welcomeMessage();
